+++ CHANGELOG.md
@@ -5,8 +5,10 @@ Full changelog for PHP Quill Renderer
 
 ## v0.80.0 
 
-* Only testing against PHP 7+. [Tests]
+* I'm now only testing against PHP 7+. [Tests]
 * Added tests for setting attributes. [Tests]
+* Switched to PSR4
+* Minor change to API if using Quill (Render) class, after the PSR4 change I didn't like Quill/Quill.
 
 ## v0.70.0 - 2017-04-19